+++ src/type_registry.rs
@@ -20,6 +20,7 @@ use crate::insert_name::{self, InsertName};
 use crate::lookup_name::{self, LookupName, LookupNameDef};
 use crate::type_shape::{self, Primitive, TypeShape, VariantDesc};
 use alloc::borrow::ToOwned;
+use alloc::format;
 use alloc::string::String;
 use alloc::vec;
 use alloc::vec::Vec;
@@ -283,6 +284,19 @@ impl TypeRegistry {
         }
 
         registry
+    }
+
+    /// Return a [`LookupName`] corresponding to one concrete type for every entry in the registry.
+    /// Any generic parameters are substituted for `()` in order that we can return valid concrete types.
+    pub fn keys(&self) -> impl Iterator<Item = LookupName> + use<'_> {
+        self.types.keys().map(|key| {
+            let insert_name = InsertName {
+                name: key.name.clone(),
+                params: (0..key.generic_params).map(|_| format!("()")).collect(),
+                pallet: key.pallet.clone(),
+            };
+            LookupName::from(insert_name)
+        })
     }
 
     /// Insert a new type into the registry.